Make CompileBot compile Java
Extremely buggy; I'm not going to try to list all of the bugs here.

The main bug is that we send a bad api request which defaults to
compiling in Java. This means that no matter what language we tell
it to compile in, it will still use Java.

I need to figure out how to send a propper api request to rextester